Fix duplicated "another gateway has already sent ack or downlink msg" Fix update fcntup for confirmed message in fcnt down race is lost
